Add simultaneous moving of plots limit Move cleaning of temp dirs to dispatcher
